Move simpler SidePanelModal implementation and utilities into kolibri-common.
