+++ packages/kolibri-common/components/courses/sidePanel/SidePanelLayout.vue
